fix: change all threshold params to double to suppress MSVC C4244
connectedpixels_u8/u16: int threshold -> double threshold
connectedpixels_u32: uint32_t threshold -> double threshold
sparse_connectedpixels_u8/u16/u32: float threshold -> double threshold
connectedpixels/sparse_connectedpixels f32 baselines: float -> double

The generated wrapper passes threshold as double (from PyFloat_AsDouble).
On MSVC, passing double to int triggers C4244 warning.  Using double
in the sig means the wrapper passes double -> double (no warning),
and the explicit cast to the target type happens in our own code
where MSVC accepts it.

+++ lib/functions/connectedpixels/connectedpixels_u32.cpp
@@ -8,13 +8,13 @@
  *  "checks": ["data.ndim == 2",
  *         "data.slow_axis == 0", "( labels.format == 'i' or labels.format == 'l' )", "labels.n == data.n"],
  *  "c_overloads": [{"when": "(data.format == 'I' or data.format == 'L') and data.itemsize == 4 and data.ndim == 2 and data.slow_axis == 0",
- *         "sig": "int connectedpixels_u32(const uint32_t *data, int32_t *labels, uint32_t threshold, int verbose, int eightconnected, intptr_t ns, intptr_t nf) -> int",
+  *         "sig": "int connectedpixels_u32(const uint32_t *data, int32_t *labels, double threshold, int verbose, int eightconnected, intptr_t ns, intptr_t nf) -> int",
  *      "map": {"data": "data.ptr", "labels": "labels.ptr", "threshold": "threshold", "verbose": "verbose", "eightconnected": "con8", "ns": "data.shape[0]", "nf": "data.shape[1]"}}]}
 C2PY_END */
 
 extern "C" int connectedpixels_u32(const uint32_t *data, int32_t *labels,
-                                    uint32_t threshold, int verbose,
-                                    int eightconnected, intptr_t ns, intptr_t nf) {
-    return connectedpixels_impl<uint32_t>(data, labels, threshold,
+                                     double threshold, int verbose,
+                                     int eightconnected, intptr_t ns, intptr_t nf) {
+    return connectedpixels_impl<uint32_t>(data, labels, (uint32_t)threshold,
                                            verbose, eightconnected, ns, nf);
 }
